pyproject.toml
[project]
name = "acc-agent"
version = "0.1.0"
description = "Agent Cognitive Compressor (ACC) Implementation"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "langchain",
    "langchain-openai",
    "python-dotenv",
    "chromadb",
    "langchain-community",
    "langchain-google-genai"
]
requires-python = ">=3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/acc_agent"]
